requirements.txt
PyYAML==6.0.2
pandas==2.2.3
scikit-learn==1.6.1
tensorflow==2.19.0
matplotlib==3.10.1